Move TokenizeJson to NJsonIndex, suppport many tokens, support subpath and empty path
