+++ ydb/core/base/ut/fulltext_ut.cpp
@@ -7,65 +7,6 @@ namespace NKikimr::NFulltext {
 
 Y_UNIT_TEST_SUITE(NFulltext) {
 
-    Y_UNIT_TEST(TokenizeJson) {
-        TString error;
-
-        UNIT_ASSERT_VALUES_EQUAL(TokenizeJson("\"invalid json", error), TVector<TString>{});
-        UNIT_ASSERT(!error.empty());
-
-        UNIT_ASSERT_VALUES_EQUAL(TokenizeJson("\"literal string\"", error), TVector<TString>{TString("\0\3literal string", 16)});
-        UNIT_ASSERT_VALUES_EQUAL(error, "");
-
-        TString obj = "{\"id\":42042,\"brand\":\"bricks\",\"part_count\":1401,\"price\":null,\"parts\":"
-            "[{\"id\":32526,\"count\":7,\"name\":\"3x5\"},{\"id\":32523,\"count\":17,\"name\":\"1x3\"}]}";
-        auto tokens = TokenizeJson(obj, error);
-        std::sort(tokens.begin(), tokens.end());
-        UNIT_ASSERT_VALUES_EQUAL(tokens, (TVector<TString>{
-            TString("\2id", 3),
-            TString("\2id\0\4\0\0\0\0@\x87\xE4@", 13),
-            TString("\5brand", 6),
-            TString("\5brand\0\3bricks", 14),
-            TString("\5parts", 6),
-            TString("\5parts\2id", 9),
-            TString("\5parts\2id", 9),
-            TString("\5parts\2id\0\4\0\0\0\0\x80\xC3\xDF@", 19),
-            TString("\5parts\2id\0\4\0\0\0\0\xC0\xC2\xDF@", 19),
-            TString("\5parts\4name", 11),
-            TString("\5parts\4name", 11),
-            TString("\5parts\4name\0\0031x3", 16),
-            TString("\5parts\4name\0\0033x5", 16),
-            TString("\5parts\5count", 12),
-            TString("\5parts\5count", 12),
-            TString("\5parts\5count\0\4\0\0\0\0\0\0\x1C@", 22),
-            TString("\5parts\5count\0\4\0\0\0\0\0\0001@", 22),
-            TString("\5price", 6),
-            TString("\5price\0\2", 8),
-            TString("\npart_count", 11),
-            TString("\npart_count\0\4\0\0\0\0\0\xE4\x95@", 21)
-        }));
-        UNIT_ASSERT_VALUES_EQUAL(error, "");
-
-        TString emptyKeyObj = "{\"\":{\"a\":\"b\"}}";
-        UNIT_ASSERT_VALUES_EQUAL(TokenizeJson(emptyKeyObj, error), (TVector<TString>{
-            TString("\0", 1),
-            TString("\0\1a", 3),
-            TString("\0\1a\0\3b", 6)
-        }));
-        UNIT_ASSERT_VALUES_EQUAL(error, "");
-
-        TString longKey;
-        longKey.resize(1000);
-        for (size_t i = 0; i < longKey.size(); i++)
-            longKey[i] = 'a';
-        TString longKeyObj = "{\"" + longKey + "\":{\"short\":\"b\"}}";
-        UNIT_ASSERT_VALUES_EQUAL(TokenizeJson(longKeyObj, error), (TVector<TString>{
-            TString("\xE8\7") + longKey,
-            TString("\xE8\7") + longKey + TString("\5short", 6),
-            TString("\xE8\7") + longKey + TString("\5short\0\3b", 9)
-        }));
-        UNIT_ASSERT_VALUES_EQUAL(error, "");
-    }
-
     Y_UNIT_TEST(ValidateColumnsMatches) {
         TString error;
