+++ ydb/core/tx/datashard/build_index/fulltext.cpp
@@ -7,6 +7,7 @@
 #include <ydb/core/base/appdata.h>
 #include <ydb/core/base/counters.h>
 #include <ydb/core/base/fulltext.h>
+#include <ydb/core/base/json_index.h>
 #include <ydb/core/kqp/common/kqp_types.h>
 #include <ydb/core/scheme/scheme_tablecell.h>
 
@@ -204,10 +205,10 @@ public:
         TVector<TString> tokens;
         if (Request.GetIndexType() == NKikimrTxDataShard::EFulltextIndexType::Json) {
             if (IsBinaryJson) {
-                tokens = TokenizeBinaryJson(row.Get(0).AsBuf());
+                tokens = NJsonIndex::TokenizeBinaryJson(row.Get(0).AsBuf());
             } else {
                 TString error;
-                tokens = TokenizeJson(row.Get(0).AsBuf(), error);
+                tokens = NJsonIndex::TokenizeJson(row.Get(0).AsBuf(), error);
                 if (error != "") {
                     JsonErrors++;
                 }
